third_party/WebKit/LayoutTests/inspector/elements/styles-1/edit-resource-referred-by-multiple-styletags-expected.txt

Tests that editing sourcecode which is referred by multiple stylesheets (via sourceURL comment) updates all stylesheets.

Inspected node
[expanded] 
element.style { ()

[expanded] 
#inspected { (stylesheet.css:2 -> stylesheet.css:2:13)
    font-size: 2em;

[expanded] 
#inspected { (stylesheet.css:2 -> stylesheet.css:2:13)
/-- overloaded --/     font-size: 2em;

[expanded] 
#inspected { (stylesheet.css:2 -> stylesheet.css:2:13)
/-- overloaded --/     font-size: 2em;

[expanded] 
div { (user agent stylesheet)
    display: block;
